[python] Move `options` implementations modules to private.
This sets the sub-modules of `options` to be "private". The name
`SOMATileDBContext` remains exported, but internal-only things are now
explicitly internal-only.
