+++ apis/python/src/tiledbsoma/options/__init__.py
@@ -1,4 +1,4 @@
-from .soma_tiledb_context import SOMATileDBContext
+from ._soma_tiledb_context import SOMATileDBContext
 
 __all__ = [
     "SOMATileDBContext",
